# Conflicts:
#   app/inventory.py
#   app/templates/addinventory.html

+++ app/models/user.py
@@ -15,6 +15,7 @@ class User(UserMixin):
         self.seller = seller
         self.balance = "{:0.2f}".format(balance)
 
+    #method for finding a user given an email and password. Returns a User object
     @staticmethod
     def get_by_auth(email, password):
         rows = app.db.execute("""
@@ -31,6 +32,7 @@ WHERE email = :email
         else:
             return User(*(rows[0][1:]))
 
+    #returns a boolean if an email is already in the Users table
     @staticmethod
     def email_exists(email):
         rows = app.db.execute("""
@@ -41,6 +43,7 @@ WHERE email = :email
                               email=email)
         return len(rows) > 0
 
+    #adds a user to the Users table. returns a User object
     @staticmethod
     def register(email, password, firstname, lastname, address):
         try:
@@ -60,6 +63,7 @@ WHERE email = :email
             print(str(e))
             return None
 
+    #returns a User object given their user id
     @staticmethod
     @login.user_loader
     def get(id):
@@ -71,8 +75,10 @@ WHERE id = :id
                               id=id)
         return User(*(rows[0])) if rows else None
 
+    #update a user's personal information. Returns an updated User object
     @staticmethod
     def updateUser(id, newEmail, newPassword, newAddress, newFirstName, newLastName, newBalance, passwordConfirmation):
+        #get the user's old information
         rows = app.db.execute("""
             SELECT password, email, address, firstname, lastname, balance
             FROM Users
@@ -86,11 +92,14 @@ WHERE id = :id
         oldlastname = rows[0][4]
         oldbalance = rows[0][5] 
 
+        #check if a user has entered the correct password. if the PW is incorrect, the update is aborted
         if not check_password_hash(oldpassword, passwordConfirmation):
             # incorrect password
             raise BadUpdateException("Incorrect Password")
 
-
+        #check that the new email is:
+        #   - not the same as the old email
+        #   - not an email that exists in the database for a different user
         if len(newEmail) > 0:
             if newEmail == oldemail:
                 raise BadUpdateException("New email can't be old email")
@@ -98,25 +107,41 @@ WHERE id = :id
                 raise BadUpdateException("Email already exists")
             app.db.execute("UPDATE Users SET email = :newEmail WHERE id = :id",id=id, newEmail=newEmail)
 
+        #check that the new password is:
+        #   - not the same as the old password
         if len(newPassword) > 0:
             if check_password_hash(oldpassword, newPassword):
                 raise BadUpdateException("New password can't be old password")
             app.db.execute("UPDATE Users SET password = :newPShash WHERE id = :id",id=id, newPShash=generate_password_hash(newPassword))
+        
+        #check that the new address is:
+        #   - not the same as the old address
         if len(newAddress) > 0:
             if oldaddress == newAddress:
                 raise BadUpdateException("New address can't be old address")
             app.db.execute("UPDATE Users SET address = :newAddress WHERE id = :id",id=id, newAddress=newAddress)
+        
+        #check that the new first name is:
+        #   - not the same as the old first name
         if len(newFirstName) > 0:
             if oldfirstname == newFirstName:
                 raise BadUpdateException("New first name can't be old first name")
             app.db.execute("UPDATE Users SET firstname = :newFirstName WHERE id = :id",id=id, newFirstName=newFirstName)
+        
+        #check that the new last name is:
+        #   - not the same as the old last name
         if len(newLastName) > 0:
             if oldlastname == newLastName:
                 raise BadUpdateException("New last name can't be old last name")
             app.db.execute("UPDATE Users SET lastname = :newLastName WHERE id = :id",id=id, newLastName=newLastName)
+        
+        #check that the new balance is:
+        #   - a valid float
+        #   - a positive number
+
         if len(newBalance) > 0:
             try:
-                newBalance = int(newBalance)
+                newBalance = float(newBalance)
             except Exception:
                 raise BadUpdateException("New balance must be a number")
 
@@ -126,6 +151,7 @@ WHERE id = :id
 
         return User.get(id)
 
+#exception for when a user tries to update their information and fails
 class BadUpdateException(BaseException):
     def __init__(self, msg):
         super().__init__()
